Added raw in model receivers

+++ ViaSofie/dossiers/models.py
@@ -65,6 +65,8 @@ def auto_delete_file_on_delete(sender, instance, **kwargs):
     """Deletes file from filesystem
     when corresponding `MediaFile` object is deleted.
     """
+    if kwargs.get('raw'):
+        return
     if instance.foto:
         if os.path.isfile(instance.foto.path):
             os.remove(instance.foto.path)
@@ -75,6 +77,8 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
     """Deletes file from filesystem
     when corresponding `MediaFile` object is changed.
     """
+    if kwargs.get('raw'):
+        return
     if not instance.pk:
         return False
